Fixed some issues with the new array

+++ amulet/world_interface/chunk/interfaces/leveldb/base_leveldb_interface.py
@@ -330,7 +330,7 @@ class BaseLevelDBInterface(Interface):
                         return_inverse=True,
                         axis=0,
                     )
-                    blocks[cy] = sub_chunk_blocks.reshape(16, 16, 16) + len(palette)
+                    blocks[cy] = sub_chunk_blocks.reshape(16, 16, 16).astype(numpy.uint32) + len(palette)
                     palette += [
                         tuple(
                             sub_chunk_palette[storage_index][index]
